+++ tests/test_lock_manager.py
@@ -9,9 +9,9 @@ import unittest
 from tempfile import gettempdir
 from time import sleep
 
-from wsgidav import lock_manager, lock_storage
 from wsgidav.dav_error import DAVError
-from wsgidav.lock_storage_redis import LockStorageRedis
+from wsgidav.lock_man import lock_manager, lock_storage
+from wsgidav.lock_man.lock_storage_redis import LockStorageRedis
 
 # ========================================================================
 # BasicTest