+++ wsgidav/lock_man/lock_manager.py
@@ -9,8 +9,8 @@ Implements the `LockManager` object that provides the locking functionality.
 The LockManager requires a LockStorage object to implement persistence.
 Two alternative lock storage classes are defined in the lock_storage module:
 
-- wsgidav.lock_storage.LockStorageDict
-- wsgidav.lock_storage.LockStorageShelve
+- wsgidav.lock_man.lock_storage.LockStorageDict
+- wsgidav.lock_man.lock_storage.LockStorageShelve
 
 
 The lock data model is a dictionary with these fields: