+++ wsgidav/lock_man/lock_storage_redis.py
@@ -9,7 +9,7 @@ import time
 import redis
 
 from wsgidav import util
-from wsgidav.lock_manager import (
+from wsgidav.lock_man.lock_manager import (
     generate_lock_token,
     lock_string,
     normalize_lock_root,